requirements.txt
# === 台股資料 ===
FinMind>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0

# === 技術分析 ===
ta>=0.11.0

# === Web Framework ===
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0
python-multipart>=0.0.6

# === HTTP Requests ===
httpx>=0.25.0
aiohttp>=3.9.0

# === Scheduling ===
apscheduler>=3.10.4

# === Utilities ===
python-dotenv>=1.0.0
pyyaml>=6.0.1

# === Logs ===
loguru>=0.7.0

# === Phase 2 News ===
feedparser>=6.0.10
beautifulsoup4>=4.12.3

# === AI 分析 ===
google-generativeai>=0.3.0

# === 資料庫 ===
SQLAlchemy>=2.0.0
